update dependencies and fix failing test (#5)
Correctly handle the case of an empty bzip2 file. The original code worked with go1.13 but now go1.14 and on because
of bug in io.Pipe which could cause errors to be lost. The bug in io.Pipe was fixed in https://github.com/golang/go/commit/f45eb9ff3c96dfd951c65d112d033ed7b5e02432 which exposed the error in pbzip2.

Also add CI/linting.