handle false positive detection of bzip2 magic numbers in bzip2 data (#8)
As explained in the README the scanner used to detect bzip2 blocks can be fooled by false positive detection of the block magic number occurring naturally in the compressed data. This change attempts to merge bzip2 blocks if one fails to decompress, if successful, then it can be safely assumed that the blocks were split by such a false positive. This reduces the probability of failing to decompress a file because of a false positive from P(magic occurring in unlimited data) to P(magic occurring twice in one bzip block).