fix a bug in the bitstream.Scan. (#16)
As pointed out by #13 he bitstream scanner would call log.Fatal if it detected a 'gap' between the a match on the first 4 bytes of the magic # and the last 6 bytes. This PR addresses this by correctly catching this case as a failure to match the magic #. See TestScanInconsistency for how to repro this case.