fix for #33 (#34)
Issue #33 reports a hang when processing a corrupted input file. The hang is due to the logic in tryMergeBlocks that wait indefinitely for the 'next block' even one does not exist and the input stream has been completely processed. This occurs because there is not test for the input channel being closed unless the heap of decompressed blocks is empty. The fix is simply to check for the channel being closed even when the heap is not empty.